requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
blas=1.0=mkl
ca-certificates=2019.11.28=hecc5488_0
certifi=2019.11.28=py36_0
cffi=1.13.2=py36h7a1dbc1_0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=py_2
freetype=2.9.1=ha9979f8_1
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
intel-openmp=2019.4=245
jpeg=9b=hb83a4c4_2
kiwisolver=1.1.0=py36he980bc4_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_0
matplotlib=3.1.2=py36_1
matplotlib-base=3.1.2=py36h2981e6d_1
mkl=2019.4=245
mkl-service=2.3.0=py36hb782905_0
mkl_fft=1.0.15=py36h14836fe_0
mkl_random=1.1.0=py36h675688f_0
ninja=1.9.0=py36h74a9793_0
numpy=1.17.4=py36h4320e6b_0
numpy-base=1.17.4=py36hc3f5095_0
olefile=0.46=py36_0
openssl=1.1.1d=hfa6e2cd_0
pillow=6.2.1=py36hdc69c19_0
pip=19.3.1=py36_0
pycparser=2.19=py36_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py36h6538335_4
python=3.6.9=h5500b2f_0
python-dateutil=2.8.1=py_0
pytorch=1.3.1=py3.6_cuda101_cudnn7_0
qt=5.9.7=vc14h73c81de_0
setuptools=42.0.2=py36_0
sip=4.19.8=py36h6538335_0
six=1.13.0=py36_0
sqlite=3.30.1=he774522_0
tk=8.6.8=hfa6e2cd_0
torchvision=0.4.2=py36_cu101
tornado=6.0.3=py36hfa6e2cd_0
tqdm=4.41.0=py_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
wheel=0.33.6=py36_0
wincertstore=0.2=py36h7fe50ca_0
xz=5.2.4=h2fa13f4_4
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
